Inline MTL-OFI

Conflicts:
	ompi/mca/mtl/ofi/mtl_ofi_recv.c
