+++ ompi/mca/mtl/ofi/mtl_ofi_component.c
@@ -416,82 +416,4 @@ error:
 }
 
 
-int
-ompi_mtl_ofi_get_error(int error_num)
-{
-    int ret;
 
-    switch (error_num) {
-    case 0:
-        ret = OMPI_SUCCESS;
-        break;
-    default:
-        ret = OMPI_ERROR;
-    }
-
-    return ret;
-}
-
-
-int
-ompi_mtl_ofi_progress(void)
-{
-    int ret, count = 0;
-    struct fi_cq_tagged_entry wc;
-    struct fi_cq_err_entry error;
-    ompi_mtl_ofi_request_t *ofi_req = NULL;
-
-    /**
-     * Read the work completions from the CQ.
-     * From the completion's op_context, we get the associated OFI request.
-     * Call the request's callback.
-     */
-    while (true) {
-        memset(&wc, 0, sizeof(wc));
-        ret = fi_cq_read(ompi_mtl_ofi.cq, (void *)&wc, 1);
-        if (ret > 0) {
-            count++;
-            if (NULL != wc.op_context) {
-                ofi_req = TO_OFI_REQ(wc.op_context);
-                assert(ofi_req);
-                ret = ofi_req->event_callback(&wc, ofi_req);
-                if (OMPI_SUCCESS != ret) {
-                    opal_output(ompi_mtl_base_framework.framework_output,
-                                "Error returned by request event callback: %d",
-                                ret);
-                    abort();
-                }
-            }
-        } else if (ret == -FI_EAVAIL) {
-            /**
-             * An error occured and is being reported via the CQ.
-             * Read the error and forward it to the upper layer.
-             */
-            memset(&error, 0, sizeof(error));
-            ret = fi_cq_readerr(ompi_mtl_ofi.cq,
-                                &error,
-                                0);
-            if (ret) {
-                opal_output(ompi_mtl_base_framework.framework_output,
-                            "Error returned from fi_cq_readerr: %d", ret);
-            }
-
-            assert(error.op_context);
-            ofi_req = TO_OFI_REQ(error.op_context);
-            assert(ofi_req);
-            ret = ofi_req->error_callback(&error, ofi_req);
-            if (OMPI_SUCCESS != ret) {
-                opal_output(ompi_mtl_base_framework.framework_output,
-                        "Error returned by request error callback: %d",
-                        ret);
-                abort();
-            }
-        } else {
-            /**
-             * The CQ is empty. Return.
-             */
-            break;
-        }
-    }
-    return count;
-}
